Fix cython compiled file dependency generation.
The CMake find_file command usually caches a file location so it can be set by
the user and so redundant searches do not occur.  However, if a file is found in
this case, we still want to perform the search every time.  So, we have to call
the CMake unset command on the variable used.